+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+NOTES:
+
+* Update network pool server response payload
+
 ## 0.4.3 (August 16, 2024)
 
 NOTES: